migrate testsuite to pytest
nose is deprecated and abandoned and doesn't work on modern python
versions. It isn't packaged by distros and installing it successfully
installs a broken "nosetests" command that doesn't work with python 3.10
and on.

pytest has automatic support for sugaring asserts without special
syntax, and automatic support for not treating non-test functions as
tests. It works great and without any effort. Let's use it. :)

Signed-off-by: Eli Schwartz <eschwartz@gentoo.org>