Revert ":fire: Delete MANIFEST file"
This reverts commit 6b51403cf2213dc222c4532f1573c4326e63468e.

The commit never made sense. It deleted MANIFEST, but a couple years
later MANIFEST.in got added right back, which included the same files
MANIFEST did. And the commit added a data_files argument to setup.py
which was totally invalid -- it tried to install README.rst as
/usr/README.rst which was clearly wrong.

The original intention appears to have been to borrow the side effect of
data_files, and have the files appear in the sdist so setup.py could
read them, but it was the wrong way to do it.

Fixes: https://github.com/ikegami-yukino/jaconv/issues/28