port setuptools to setup.cfg
setup.cfg can be freely mixed with setup.py and settings from both are
merged together. Using the cfg file, it is possible to encode most
metadata statically, and use a few powerful new tools in the process --
in particular, it is no longer necessary to regex the `__init__.py` file
for the version, since setuptools can natively extract the relevant
property (and yes, it does so without importing the module).